+++ src/nemesis-proto_ip.c
@@ -65,7 +65,7 @@ int buildip(ETHERhdr *eth, IPhdr *ip, FileData *pd, FileData *ipod, libnet_t *l)
 	if (verbose == 3)
 		nemesis_hexdump(pkt, ip_packetlen, HEX_RAW_DECODE);
 
-	if (n != ip_packetlen) {
+	if (n != (int)ip_packetlen) {
 		fprintf(stderr, "ERROR: Incomplete packet injection.  Only wrote %d bytes.\n", n);
 	} else {
 		if (verbose) {
